altered genotype to add phenotyping center
after email discussion with Terry, we needed to include the phenotyping center as part of the genotype, since independent mice were generated from the same ES clones at separate places.  to do this, the background strains were augmented with the phenotyping center, (like BL/6 -> BL/6/WSTI) to be an instance of BL/6.  this is probably not official mouse nomenclature (for labels), but gets the point across.