Create incoming_curie_map.yaml
This is a woefully incomplete start at what classes and sources of entities are served as pages within the monarch app. An attempt to see if this kind of a file would be useful to address https://github.com/monarch-initiative/monarch-app/issues/1201. Broadly speaking, I would like to see this fleshed out for knowledge beacon purposes and get Monarch integrated as a 3rd party provider within n2t/identifiers.org registry.